73 funcionalidad de eliminar diapositivas

+++ python-app/lib/pivot_table/recalculate.py
@@ -2,14 +2,14 @@ from lib.pivot_table.get_combinable_filters import get_combinable_filters
 from lib.pivot_table.ordered_filters import ordered_filters
 from lib.pivot_table.get_data_of_frame import get_data_of_frame
 from lib.pivot_table.ordered_filters import find_filter
-from lib.pivot_table.plot_pivot_table import plot_from_entities
+from lib.pivot_table.render_bare_preview_of_pivot_table import render_bare_preview_of_pivot_table
 from lib.data_filter.is_valid_filter import is_valid_filter
 from lib.data_frame.data_frame_io import import_data_frame
-from lib.directory_definitions import bare_preview_of_pivot_table
+from lib.slide.render_preview import render_preview
+from lib.slide.delete_preview import delete_preview
 
 from models.pivot_table.self import PivotTable
 from models.pivot_table.data_filter.self import DataFilter
-from models.pivot_table.pivot_table_level import PivotTableLevel
 from models.pivot_table.data_filter.charting_mode import ChartingMode
 from models.report.self import Report
 
@@ -39,7 +39,7 @@ def fix_charts(filters: list[DataFilter], filters_order: list[PivotTable]) -> No
         new_chart.charting_mode = ChartingMode.CHART
     
 
-def recalculate(report: Report, pivot_table: PivotTable, preloaded_data_frame: pandas.DataFrame | None = None):
+def recalculate(root_directory: str, pivot_table: PivotTable, preloaded_data_frame: pandas.DataFrame | None = None):
     '''
     Re calculates the data of the pivot table based on its filters. It modifies the pivot table in place.
     Also, it creates a new preview
@@ -60,11 +60,12 @@ def recalculate(report: Report, pivot_table: PivotTable, preloaded_data_frame: p
         aggregate_function=pivot_table.aggregate_function
         )
     
-    if os.path.exists(pivot_table.bare_preview):
-        os.remove(pivot_table.bare_preview)
-    
-    filepath = bare_preview_of_pivot_table(root_directory=report.root_directory, slide_id=pivot_table.identifier)
-    plot_from_entities(filepath=filepath, pivot_table=pivot_table)
-
     pivot_table.data = new_data
-    pivot_table.bare_preview = filepath
+    
+    if pivot_table.bare_preview is not None and os.path.exists(pivot_table.bare_preview):
+        os.remove(pivot_table.bare_preview)
+
+    render_bare_preview_of_pivot_table(root_directory=root_directory, pivot_table=pivot_table)
+
+    delete_preview(slide=pivot_table)
+    render_preview(root_directory=root_directory, slides=pivot_table)
